+++ lgn-provers/src/provers/v1/query/mod.rs
@@ -1,19 +1,12 @@
 use std::collections::HashMap;
 
-use tracing::debug;
 use tracing::info;
 
-use crate::provers::v1::query::prover::StorageQueryProver;
-use crate::provers::v1::query::task::Querying;
+use crate::provers::LgnProver;
 
-pub(crate) mod prover;
-pub mod task;
-
-#[cfg(feature = "dummy-prover")]
 pub(crate) mod dummy_prover;
-
-#[cfg(not(feature = "dummy-prover"))]
 pub(crate) mod euclid_prover;
+pub use euclid_prover::EuclidQueryProver;
 
 pub const ROW_TREE_MAX_DEPTH: usize = 25;
 pub const INDEX_TREE_MAX_DEPTH: usize = 26;
@@ -31,26 +24,22 @@ pub fn create_prover(
     dir: &str,
     file: &str,
     checksums: &HashMap<String, blake3::Hash>,
-) -> anyhow::Result<Querying<impl StorageQueryProver>> {
+) -> anyhow::Result<impl LgnProver> {
+    #[cfg(feature = "dummy-prover")]
     let prover = {
-        #[cfg(feature = "dummy-prover")]
-        let prover = {
-            use dummy_prover::DummyProver;
-            info!("Creating dummy query prover");
-            DummyProver
-        };
-
-        #[cfg(not(feature = "dummy-prover"))]
-        let prover = {
-            info!("Creating query prover");
-
-            euclid_prover::EuclidQueryProver::init(url, dir, file, checksums)?
-        };
-
-        debug!("Query prover created");
-
-        prover
+        use dummy_prover::DummyProver;
+        info!("Creating dummy query prover");
+        DummyProver
     };
 
-    Ok(Querying::new(prover))
+    #[cfg(not(feature = "dummy-prover"))]
+    let prover = {
+        info!("Creating query prover");
+
+        euclid_prover::EuclidQueryProver::init(url, dir, file, checksums)?
+    };
+
+    info!("Query prover created");
+
+    Ok(prover)
 }
